refactor(classes): rinominata directory e namespace
