+++ APL-Project-Client/Properties/DataSources/APL_Project_Client.Model.Ferie.datasource
@@ -6,5 +6,5 @@
     cause the file to be unrecognizable by the program.
 -->
 <GenericObjectDataSource DisplayName="Ferie" Version="1.0" xmlns="urn:schemas-microsoft-com:xml-msdatasource">
-  <TypeInfo>APL_Project_Client.Model.Ferie, APL-Project-Client, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null</TypeInfo>
+  <TypeInfo>APL_Project_Client.Classes.Ferie, APL-Project-Client, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null</TypeInfo>
 </GenericObjectDataSource>
